+++ Starstorm 2/Cores/BuffCore.cs
@@ -28,6 +28,7 @@ namespace Starstorm2Unofficial.Cores
 
         public static BuffDef exeSuperchargedBuff;
         public static BuffDef nucleatorSpecialBuff;
+        public static BuffDef nucleatorSpecialDebuff;
 
         public static BuffDef chirrFriendBuff;
         public static BuffDef chirrSelfBuff;
@@ -50,6 +51,7 @@ namespace Starstorm2Unofficial.Cores
             //LogCore.LogInfo("Initializing Core: " + base.ToString());
             exeSuperchargedBuff = CreateBuffDef("SS2UExecutionerSuperchargedBuff", false, false, false, new Color(72 / 255, 1, 1), LegacyResourcesAPI.Load<Sprite>("Textures/BuffIcons/texBuffNullifiedIcon"));
             nucleatorSpecialBuff = CreateBuffDef("SS2UNucleatorSpecialBuff", true, false, false, new Color32(224, 217, 67, 255), Modules.Assets.mainAssetBundle.LoadAsset<Sprite>("radiation_debuff"));
+            nucleatorSpecialDebuff = CreateBuffDef("SS2UNucleatorSpecialDebuff", false, false, false, new Color32(224, 217, 67, 255), Modules.Assets.mainAssetBundle.LoadAsset<Sprite>("radiation_debuff"));
 
             detritiveBuff = ScriptableObject.CreateInstance<BuffDef>();
             detritiveBuff.buffColor = Color.white;
@@ -185,7 +187,7 @@ namespace Starstorm2Unofficial.Cores
                 }
                 else
                 {
-                    Debug.LogError("Starstorm 2 Unofficial: Failed to set up Chirr Friend Buff overlay IL Hook.");
+                    Debug.LogError("Starstorm 2 Unofficial: Failed to set up Weak overlay IL Hook.");
                 }
 
                 c = new ILCursor(il);
@@ -202,7 +204,7 @@ namespace Starstorm2Unofficial.Cores
                 }
                 else
                 {
-                    Debug.LogError("Starstorm 2 Unofficial: Failed to set up Chirr Friend Distract Buff overlay IL Hook.");
+                    Debug.LogError("Starstorm 2 Unofficial: Failed to set up FullCrit overlay IL Hook.");
                 }
             };